+++ pleroma_bot/tests/test_logger.py
@@ -107,7 +107,7 @@ def test_post_pleroma_media_logger(rootdir, sample_users, caplog):
                 mock.post(media_url, status_code=413)
                 with caplog.at_level(logging.ERROR):
                     sample_user_obj.post_pleroma(
-                        (test_user.pinned, ""), None, False
+                        (test_user.pinned, "", ""), None, False
                     )
                 assert 'Exception occurred' in caplog.text
                 assert 'Media size too large' in caplog.text
@@ -117,7 +117,7 @@ def test_post_pleroma_media_logger(rootdir, sample_users, caplog):
                         requests.exceptions.HTTPError
                 ) as error_info:
                     sample_user_obj.post_pleroma(
-                        (test_user.pinned, ""), None, False
+                        (test_user.pinned, "", ""), None, False
                     )
                 exception_value = (
                     f"500 Server Error: None for url: {media_url}"